Fix dual-nic bare-metal network issue
If the default route was on the the second interface due to
the way Debian's default network configuration treated the
dual-nic hardware, then without this fix the network would
fail when the second interface was moved to the bridge.

We now restart the system's network after configuring the
bridge on the second interface.

Signed-off-by: Daniel F. Dickinson <dfdpublic@wildtechgarden.ca>