Unify and fix port forwarding from external to VMs
Only the first (lexicographically) hook was getting applied in the
previous configuration, so we update the handling of the split
port-forward hooks to ensure they are all applied.

In addition were unify the handling of 'early' (before we can
gather facts due to lack of SSH path into the virtual host)
hooks and the main batch of hooks (for which we can determine
their IP addresses via gathered facts from the instances
themselves). This avoids duplication and ensure the same
behaviour for all hooks.

Signed-off-by: Daniel F. Dickinson <dfdpublic@wildtechgarden.ca>