Refactor HAProxy and make imap a port forward instead
Since we are port forwarding to the HAProxy instance as well
as to the dovecot (IMAP) instead it doesn't make sense to have
a 'double-jump' by doing IMAP through HAProxy. Therefore make
IMAP a directory port forward.

While we are at it, refactor HAProxy to use a configuration
directory with multiple files instead of a single file config.

Signed-off-by: Daniel F. Dickinson <dfdpublic@wildtechgarden.ca>